ARM: dts: pico-imx8mq: enable 5 inch ILI9881C MIPI DSI panel
This patch improves the 'success rate' a lot.
'success rate' is defined as 'when the board boots up, it's able to see clear tux picture.'
Before applying this patch, the success rate is only 5% (1/20 boots).
But it's still defective.
It has chance to show white and flickering display on screen.
'failure rate' is 15% (9/60 boots)